Holy fucking shit it actually works, what the fuck?

+++ include/vector2-ext.hpp
@@ -4,10 +4,10 @@
 
 namespace Vector2Ext {
 float Distance(Vector2 a, Vector2 b);
-int SquaredSize(Vector2 a);
+float SquaredSize(Vector2 a);
 Vector2 Difference(Vector2 a, Vector2 b);
 Vector2 Add(Vector2 a, Vector2 b);
-Vector2 Multiply(Vector2 a, int b);
-int DistanceSquared(Vector2 a, Vector2 b);
-int Dot(Vector2 a, Vector2 b);
+Vector2 Multiply(Vector2 a, float b);
+float DistanceSquared(Vector2 a, Vector2 b);
+float Dot(Vector2 a, Vector2 b);
 }; // namespace Vector2Ext
